fix(38988552): rename util name
